% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{adaptivesoft.thre}
\alias{adaptivesoft.thre}
\title{adaptivesoft.thre() Function}
\usage{
adaptivesoft.thre(G, phi, eta = 2, diagTF = TRUE)
}
\arguments{
\item{G, }{phi, diagTF}
}
\description{
Adaptive soft-thresholding complex matrix G
}
\examples{
adaptivesoft.thre(G)




}
\keyword{soft.thre}
