% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{chol.complex}
\alias{chol.complex}
\title{Cholesky decomposition of complex maxtrix}
\usage{
\method{chol}{complex}(G)
}
\description{
chol.complex() Function
}
\examples{
chol.complex(G)



}
\keyword{(complex)}
\keyword{Cholesky}
\keyword{decomposition}
