% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mLRDglasso.R
\name{clime.complex}
\alias{clime.complex}
\title{Sparse estimation of inverse G (precison matrix) using LW CLIME method.}
\usage{
clime.complex(
  Ghat,
  Tt,
  lambda = "ebic",
  type = "soft",
  gridTF = TRUE,
  gg = 1,
  bound = c(0.05, 1),
  debiasTF = FALSE
)
}
\arguments{
\item{Ghat}{Estimated (nonsparse) long-run variance matrix to be sparsely estimated using LW-CLIME}

\item{Tt}{Data length, used in calculating BIC}

\item{lambda}{"ebic" uses extended BIC criteria to select penalty parameter in LW-CLIME.
User also can provide numerical value.}

\item{type}{Types of thresholding in ADMM algorithm. "hard", "soft" and "adaptive" threshold functions are possible.}

\item{gridTF}{If TRUE, penalty parameter is searched over interval provided on bound argument. Otherwise, optim function searches optimal
lambda.}

\item{gg}{The tuning parameter in the extended BIC criteria. Default value is 1. If gg=0, it is a usual BIC.}

\item{bound}{Bound of grid search in extended BIC. Default value is (.05, 1)}

\item{debiasTF}{If TRUE, debiased by applying constrained MLE introduced in the paper.}

\item{approxTF}{If TRUE, univariate LRD parameter is used in the estimation. Otherwise, multivariate LRD parameter estimator is used.}
}
\description{
This function estimates sparse long-run variance (complex) matrix using LW-CLIME method.
}
\details{
clime.complex
}
\examples{
clime.complex(Ghat, Tt, lambda="ebic")





}
\keyword{CLIME}
\keyword{Local}
\keyword{Whittle}
\keyword{estimation}
