% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencode.R
\name{CovarFARIMA}
\alias{CovarFARIMA}
\title{Covariance function in mlw paper (21)}
\usage{
CovarFARIMA(N, d, delta, Sigmae)
}
\arguments{
\item{N}{sample size}

\item{d}{LRD parameter}

\item{delta}{delta parameter}
}
\description{
CovarFARIMA () Function
}
\examples{
CovarFARIMA(N,d,delta,Sigmae)




}
\keyword{Gaussian}
\keyword{Process}
\keyword{generation}
