% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencode.R
\name{CovoneVARFIMA}
\alias{CovoneVARFIMA}
\title{One-sided VARFIMA(0, D, 0) model in SPIE paper equation (18)}
\usage{
CovoneVARFIMA(N, d, invS)
}
\arguments{
\item{N, }{sample size}

\item{d, }{LRD parameter, ensure the elements in d satisfy -1/2<d(k)<1/2}

\item{invS}{is a sparse inverse matrix of innovations Sigma matrix}
}
\description{
CovoneVARFIMA() Function
}
\examples{
CovoneVARFIMA(N, d, invS)

}
\keyword{Gaussian}
\keyword{Process}
\keyword{generation}
