% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencode.R
\name{CovtwoVARFIMA}
\alias{CovtwoVARFIMA}
\title{two-sided VARFIMA(0, D, 0) model in SPIE paper equation (18)}
\usage{
CovtwoVARFIMA(N, d, R, X)
}
\arguments{
\item{N, }{sample size}

\item{d, }{LRD parameter, ensure the elements in d satisfy -1/2<d(k)<1/2}

\item{R, }{R is a real upper-triangular matrix representing sparsity}

\item{X, }{X is also complex valued upper-triangular matrix such that invG = (RW)^H\%*\%(RH)}
}
\description{
CovtwoVARFIMA() Function
}
\examples{
CovtwoVARFIMA(N, d, R, X)
}
\keyword{Gaussian}
\keyword{Process}
\keyword{generation}
