% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencode.R
\name{CovtwoVARFIMA2}
\alias{CovtwoVARFIMA2}
\title{two-sided VARFIMA(0, D, 0) model in SPIE paper equation (18)}
\usage{
CovtwoVARFIMA2(N, d, G)
}
\arguments{
\item{N, }{sample size}

\item{d, }{LRD parameter, ensure the elements in d satisfy -1/2<d(k)<1/2}

\item{G, }{G is long-run variance matrix such that G = Z*t(Con(Z));}
}
\description{
CovtwoVARFIMA2() Function
}
\examples{
CovtwoVARFIMA2(N, d, G)
}
\keyword{Gaussian}
\keyword{Process}
\keyword{generation}
