% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{Det.complex}
\alias{Det.complex}
\title{Det.complex() Function}
\usage{
Det.complex(G)
}
\arguments{
\item{G}{Input complex matrix}
}
\description{
Determinant of complex square matrix
}
\examples{
Det.complex(G)



}
\keyword{Determinant}
