% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{dist.Frobenius}
\alias{dist.Frobenius}
\title{dist.Frobenius() Function}
\usage{
dist.Frobenius(G, H)
}
\arguments{
\item{G, }{H}
}
\description{
Frobenious distance between two complex matrices
}
\examples{
dist.Frobenius(G, H)




}
\keyword{Frobenious}
\keyword{norm}
