% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{dist.spectral}
\alias{dist.spectral}
\title{dist.spectral() Function}
\usage{
dist.spectral(G, H)
}
\arguments{
\item{G, }{H}
}
\description{
Spectral distance between two complex matrices
}
\examples{
dist.spectral(G, H)



}
\keyword{Spectral}
\keyword{norm}
