% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencode.R
\name{GenMultivarGauss}
\alias{GenMultivarGauss}
\title{Generating Multivariate Gaussian process using circular embedding}
\usage{
GenMultivarGauss(R, N, isAsymm = TRUE)
}
\arguments{
\item{R, }{given autocovariance matrix}

\item{N, }{sample size}
}
\description{
GenMultivarGauss () Function
}
\examples{
GenMultivarGauss(R,N, isAsymm=TRUE)



}
\keyword{Gaussian}
\keyword{Process}
\keyword{generation}
