% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mLRDglasso.R
\name{glasso.complex}
\alias{glasso.complex}
\title{Sparse estimation of inverse G using graphical lasso.}
\usage{
glasso.complex(
  Ghat,
  Tt,
  lambda = "ebic",
  type = "soft",
  gridTF = TRUE,
  gg = 1,
  bound = c(0.05, 1),
  debiasTF = FALSE
)
}
\arguments{
\item{Ghat, }{Estimated Ghat to be sparsely estimated using glasso}

\item{m}{number of frequencies used in estimation}
}
\description{
glasso.complex() Function
}
\examples{
glasso.complex(Ghat, Tt, lambda="ebic")




}
\keyword{Local}
\keyword{Whittle}
\keyword{estimation}
