% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mLRDglasso.R
\name{glassoD2}
\alias{glassoD2}
\title{Sparse constrained estimation in complex Graphical model}
\usage{
glassoD2(S, R, maxIt = 100, tol = 1e-06)
}
\arguments{
\item{S, }{input complex matrix}

\item{R, }{constraint matrix}
}
\description{
glassoD2() Function
}
\examples{
glassoD2(S, R)



}
\keyword{complex}
\keyword{constrained}
\keyword{estimation}
\keyword{matrix}
\keyword{of}
