% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{hard.thre}
\alias{hard.thre}
\title{hard.thre() Function}
\usage{
hard.thre(G, phi, diagTF = TRUE)
}
\arguments{
\item{G, }{phi, diagTF}
}
\description{
Hard thresholding complex matrix G
}
\examples{
hard.thre(G)



}
\keyword{soft.thre}
