% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwestimation.R
\name{lwe}
\alias{lwe}
\title{lwe() Function}
\usage{
lwe(data, m, lowerd, upperd)
}
\arguments{
\item{data, m, }{lowerd, upperd}
}
\description{
Local Whittle estimation of LRD parameter for univariate time series
}
\examples{
lwe(data,m, lowerd=0, upperd=1)




}
\keyword{LRD}
\keyword{Local}
\keyword{Whittle}
\keyword{estimation}
\keyword{parameter}
