% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwestimation.R
\name{lweplot}
\alias{lweplot}
\title{Plot of local Whittle estimator over the number of frequencies used}
\usage{
lweplot(data, m, lowerd, upperd)
}
\arguments{
\item{data}{Input data (univariate time series)}

\item{m}{The maximum number of frequencies used in the plot. If missing, then $T^{.8}$ is used}

\item{lowerd}{the lower bound of LW object function, default = -.5}

\item{upperd}{the upper bound of LW object function, default = 1}
}
\description{
lweplot depics local Whittle estimators over the number of frequencies used
}
\details{
lweplot
}
\examples{
lweplot(data)




}
\keyword{LRD}
\keyword{Plot}
\keyword{Whittle}
\keyword{estimators}
\keyword{local}
\keyword{of}
\keyword{parameter}
