% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwestimation.R
\name{LWGhat}
\alias{LWGhat}
\title{LWGhat() Function}
\usage{
LWGhat(I, d, m, Tt)
}
\arguments{
\item{I, }{d, m, Tt}
}
\description{
Long-run variance estimation based on local Whittle estimation of multivariate time series
}
\examples{
LWGhat(I, d, m, Tt)




}
\keyword{Long-run}
\keyword{estimation}
\keyword{variance}
