% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwestimation.R
\name{lwMLRD}
\alias{lwMLRD}
\title{lwMLRD() Function}
\usage{
lwMLRD(data, m)
}
\arguments{
\item{data, }{m}
}
\description{
Local Whittle estimation of multivariate time series
}
\examples{
lwMLRD(data, m)



}
\keyword{Local}
\keyword{Whittle}
\keyword{estimation}
