% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{Periodogram}
\alias{Periodogram}
\title{Periodogram of multivariate time series}
\usage{
Periodogram(data)
}
\arguments{
\item{data}{Input data of dim*length}
}
\description{
Periodogram Function
}
\examples{
Periodogram(data)



}
\keyword{Periodogram}
