% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{soft.thre}
\alias{soft.thre}
\title{soft.thre() Function}
\usage{
soft.thre(G, phi, diagTF = TRUE)
}
\arguments{
\item{G, }{phi, diagTF}
}
\description{
softthresholding complex matrix G
}
\examples{
soft.thre(G)


}
\keyword{soft.thre}
