% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mLRDglasso.R
\name{sparseMLRD}
\alias{sparseMLRD}
\title{two-stage estimation of LRD parameter and sparse inverse G using penalization methods}
\usage{
sparseMLRD(
  data,
  m,
  lambda = "ebic",
  type = "soft",
  method = "glasso",
  approxTF = TRUE,
  gridTF = TRUE,
  gg = 1,
  bound = c(0.05, 1),
  debiasTF = TRUE
)
}
\arguments{
\item{data}{matrix of dim*Tt}

\item{m}{number of frequencies used in estimation.  If missed, m=\link{T^.8}.}

\item{lambda}{"ebic" uses extended BIC criteria to select penalty parameter in graphical lasso.
User also can provide numerical value.}

\item{type}{Types of thresholding in ADMM algorithm. "hard", "soft" and "adaptive" threshold functions are possible.}

\item{method}{Three methods are possible. If "glasso", LW grapchial lasso is used. If "spice", long-run correlation matrix is used instead of covariance matrix.
If "clime" is used, LW CLIME estimator is provided. Default is "glasso".}

\item{approxTF}{If TRUE, univariate LRD parameter is used in the estimation. Otherwise, multivariate LRD parameter estimator is used.}

\item{gridTF}{If TRUE, penalty parameter is searched over interval provided on bound argument. Otherwise, optim function searches optimal
lambda.}

\item{gg}{The tuning parameter in the extended BIC criteria. Default value is 1. If gg=0, it is a usual BIC.}

\item{bound}{Bound of grid search in extended BIC. Default value is (.05, 1)}

\item{debiasTF}{If TRUE, debiased by applying constrained MLE introduced in the paper.}
}
\description{
This function estimates sparse long-run variance (complex) matrix of multivariate LRD model
using local Whittle graphical lasso, spice (studentized version of graphical lasso) and CLIME.
}
\examples{
sparseMLRD(data, m)



}
\keyword{Local}
\keyword{Multivariate}
\keyword{Whittle}
\keyword{estimation}
\keyword{sparse}
