% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwestimation.R
\name{LWGhat}
\alias{LWGhat}
\title{Long-run variance estimation based on local Whittle estimation of multivariate time series}
\usage{
LWGhat(I, d, m, Tt)
}
\arguments{
\item{I}{Periodogram}

\item{d}{LRD parameters}

\item{m}{number of frequencies used in LW estimation of LRD parameters}

\item{Tt}{The number of sample size}
}
\description{
LWGhat
}
\section{Functions}{
\itemize{
\item \code{LWGhat()}: function provides long-run variance of multivariate LRD series for given LRD parameter d and periodograms.

}}
\examples{
LWGhat(I, d, m, Tt)




}
\keyword{Long-run}
\keyword{estimation}
\keyword{variance}
