% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mLRDglasso.R
\name{threGhat.bic}
\alias{threGhat.bic}
\title{Sparse estimation using thresholding method}
\usage{
threGhat.bic(
  data,
  m,
  lambda = "ebic",
  bound,
  approxTF = TRUE,
  adaptiveTF = TRUE,
  eta = 1,
  gg = 1,
  debiasTF = FALSE
)
}
\arguments{
\item{data, }{input data}

\item{m, }{number of frequencies used in LW estimation}

\item{lambda, }{Selection of lambda, ebic is default.}
}
\description{
threGhat.bic() Function
}
\examples{
threGhat.bic(data)




#####################################################
}
\keyword{LW}
\keyword{Thresholding}
\keyword{long-run}
\keyword{matrix}
\keyword{variance}
