% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mLRDglasso.R
\name{threGhat.cv}
\alias{threGhat.cv}
\title{Sparse estimation using thresholding method}
\usage{
threGhat.cv(data, m, bound = c(0.01, 1), eta = 1, gg = 1, debiasTF = TRUE)
}
\arguments{
\item{data}{input data}

\item{m}{number of frequencies used in estimation. If missed, $m=T^.8$.}

\item{bound}{Bound of threshold values in grid search. Default value is (.01, 1)}

\item{eta}{The parameter in the adaptive threshould function. Default is 1.}

\item{gg}{The tuning parameter in calculating extended BIC. Default value is 1. If gg=0, it is a usual BIC.}

\item{debiasTF}{If TRUE, debiased by applying constrained MLE introduced in the paper.}
}
\description{
This function sparsely estimates long-run variance matrix of multivariate LRD series using threshoulding method.
}
\details{
threGhat.cv
}
\examples{
threGhat.cv(data)




}
\keyword{LW}
\keyword{Thresholding}
\keyword{long-run}
\keyword{matrix}
\keyword{variance}
