% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{adaptivesoft.thre}
\alias{adaptivesoft.thre}
\title{Adaptive soft-thresholding complex matrix G}
\usage{
adaptivesoft.thre(G, phi, eta = 2, diagTF = TRUE)
}
\arguments{
\item{G}{Input complex matrix}

\item{phi}{Phi parameter in the adaptive threshold function.}

\item{eta}{Eta parameter in the adaptive threshould function.}

\item{diagTF}{If TRUE, apply thresholding on the diagonal entries of G}
}
\description{
Hard-thresholding of comlex matrix G with threshould phi.
}
\details{
adaptivesoft.thre
}
\examples{
adaptivesoft.thre(G)




}
\keyword{adaptivesoft.thre}
