% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{dist.Frobenius}
\alias{dist.Frobenius}
\title{Frobenious distance between two complex matrices}
\usage{
dist.Frobenius(G, H)
}
\description{
dist.Frobenius(G, H) calculates Frobenius distance between two complex matrices G and H.
}
\details{
dist.Frobenius() Function
}
\examples{
dist.Frobenius(G, H)




}
\keyword{Frobenious}
\keyword{norm}
