% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{dist.spectral}
\alias{dist.spectral}
\title{Spectral distance between two complex matrices}
\usage{
dist.spectral(G, H)
}
\description{
dist.spectral(G, H) calculates Spectral distance between two complex matrices G and H.
}
\details{
dist.spectral() Function
}
\examples{
dist.spectral(G, H)



}
\keyword{Spectral}
\keyword{norm}
