% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mLRDglasso.R
\name{glassoD2}
\alias{glassoD2}
\title{Constrained MLE for long-run variance matrix}
\usage{
glassoD2(S, R, maxIt = 100, tol = 1e-06)
}
\arguments{
\item{S}{input complex long-run variance matrix}

\item{R}{constraint matrix consisting of 1 (if nonzero) and 0 (if zero)}

\item{maxIt}{Maximum iteration of algorithm. Default is 100}

\item{tol}{Tolerance to stop iteratrion. Default is 1e-6}
}
\description{
This function provides constrained MLE for long-run variance matrix. It is used to debias the estimated
sparse long-run variance matrix using graphical Lasso.
}
\details{
glassoD2
}
\examples{
glassoD2(S, R)



}
\keyword{Constrained}
\keyword{estimation}
\keyword{long-run}
\keyword{matrix.}
\keyword{of}
\keyword{variance}
