% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{hard.thre}
\alias{hard.thre}
\title{Hard thresholding complex matrix G}
\usage{
hard.thre(G, phi, diagTF = TRUE)
}
\arguments{
\item{G}{Input complex matrix}

\item{phi}{Threshold level parameter}

\item{diagTF}{If TRUE, apply thresholding on the diagonal entries of G}
}
\description{
Hard-thresholding of comlex matrix G with threshould phi.
}
\details{
hard.thre() Function
}
\examples{
hard.thre(G)



}
\keyword{hard.thre}
