% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwestimation.R
\name{lwMLRD}
\alias{lwMLRD}
\title{Local Whittle estimation of multivariate time series}
\usage{
lwMLRD(data, m)
}
\arguments{
\item{data}{Input data (dim*length)}

\item{m}{The number of frequencies used in estimation.  If missed, $m=\link{T^.8}$.}
}
\description{
This function provides local Whittle estimation of multivariate long range dependent time series
}
\details{
lwMLRD
}
\examples{
lwMLRD(data, m)



}
\keyword{Local}
\keyword{Whittle}
\keyword{estimation}
