% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mLRDglasso.R
\name{MLRD.glasso}
\alias{MLRD.glasso}
\title{two-stage estimatin of LRD parameter and sparse G using glasso}
\usage{
MLRD.glasso(
  data,
  m,
  lambda = "ebic",
  type = "soft",
  approxTF = TRUE,
  gridTF = TRUE,
  gg = 1,
  bound = c(0.05, 1),
  debiasTF = FALSE
)
}
\arguments{
\item{data, }{matrix of dim*Tt}

\item{m}{number of frequencies used in estimation}
}
\description{
MLRD.glasso() Function
}
\examples{
MLRD.glasso(data, m)



}
\keyword{Local}
\keyword{Whittle}
\keyword{estimation}
