% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{Periodogram}
\alias{Periodogram}
\title{Periodogram Function}
\usage{
Periodogram(data)
}
\arguments{
\item{data}{Input data}
}
\description{
Periodogram of univariate time series
}
\examples{
Periodogram(data)



}
\keyword{Periodogram}
