% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunctions.R
\name{soft.thre}
\alias{soft.thre}
\title{softthresholding complex matrix G}
\usage{
soft.thre(G, phi, diagTF = TRUE)
}
\arguments{
\item{G}{Input complex matrix}

\item{phi}{Threshold level parameter}

\item{diagTF}{If TRUE, apply thresholding on the diagonal entries of G}
}
\description{
Soft-thresholding of comlex matrix G with threshould phi.
}
\details{
soft.thre
}
\examples{
soft.thre(G)


}
\keyword{function}
\keyword{soft}
\keyword{thresholding}
